test merge  'thomas' and 'merge'
