--- data.py
@@ -1,4 +1,0 @@
-#-*- coding: utf-8
-
-
-
